Refactor dpc folders
