+++ dpc/src/state/ledger_proof.rs
